Deprecate `MarkerView` and `ProgressBar`. Move Solar extension to separate file.
