+++ Sources/MapboxNavigation/StyleManager.swift
@@ -304,13 +304,3 @@ extension Date {
         return currentSecondsFromMidnight < sunriseSecondsFromMidnight || currentSecondsFromMidnight > sunsetSecondsFromMidnight
     }
 }
-
-extension Solar {
-    init?(date: Date?, coordinate: CLLocationCoordinate2D) {
-        if let date = date {
-            self.init(for: date, coordinate: coordinate)
-        } else {
-            self.init(coordinate: coordinate)
-        }
-    }
-}
